Reorganizing ascii converter code
